build: ty joins the gates, and the tree reads zero under it
No type checker ran here; the package shipped py.typed and a stub for
its compiled half and nothing read them. ty does now, pinned like
ruff (0.0.81), configured to the 3.9 floor and to the stub the build
writes under build/python, as `make typecheck` in `make ci` and in
the CI python job. The one-time burn-down is the rest of the commit,
about 110 sites, none of them a behaviour change:

The estimator base declared none of the attributes its subclasses
store from their constructors, and read `_model` as `Model | None`
at every prediction site behind a separate `_check_fitted` call. The
base now declares what a subclass supplies, and `_fitted()` returns
the model or raises the same RuntimeError, so each site narrows in
one expression. `from_file` returns the class it was called on. The
ParamsOps mixin declares the dataclass protocol its generated
subclasses satisfy, and its constructors return the calling class,
which is what interop's `from_*` already returned at runtime. The
compiled module's private `_params_from_toml` enters the stub
(INCLUDE_PRIVATE), since the package calls it. sklearn's Tags
dataclass is built through an Any-typed alias: ty cannot yet read a
slots dataclass's fields, and the tags test pins the contract.

Bench and scripts take narrowing where the code already relied on
the value (a regex match, a spec loader, an optional list), typed
locals where a dict's first key was inferred as its type, and
`int()` where a campaign knob is integral. check_standings skips a
row with no numeric metric where it would have raised on it; no
committed row is such a row. The tests' `**kwargs` dicts get the
types the constructors accept, and a script carrying PEP 723
metadata is left to `uv run` (exclude-scripts).

Ran: make typecheck (0 diagnostics), make python-test (510 passed),
make lint-python, make format-check, make params-check, make
docs-check, make site

+++ python/bonsai/bench/driver.py
@@ -192,7 +192,9 @@ class DeviceMemSampler:
             self._stop.wait(self.interval_s)
 
 
-def run_one(spec: dict, timeout: int, sampler: bool = False, data_cache: str | None = None) -> dict:
+def run_one(
+    spec: dict, timeout: float, sampler: bool = False, data_cache: str | None = None
+) -> dict:
     """One worker child, one cell: the payload dict for the row.
 
     Timeouts, signals, and nonzero exits come back as status rows
@@ -372,7 +374,7 @@ def _default_query() -> tuple:
     # Any NVML failure (missing package, no driver, MIG topology) means the
     # fallback, never a dead sampler.
     try:
-        import pynvml
+        import pynvml  # ty: ignore[unresolved-import]
 
         pynvml.nvmlInit()
         handle = pynvml.nvmlDeviceGetHandleByIndex(_device_index())
